Impl del rotate a nivel jugador + refactor player.cpp

+++ common/object_DTO.cpp
@@ -2,7 +2,7 @@
 /* Player */
 ObjectDTO::ObjectDTO(ObjectType type, const std::vector<uint16_t>& position, uint16_t id,
                      PlayerType player_type, WeaponModel weapon_model, uint16_t health,
-                     uint16_t money, uint16_t ammo, _Float16 angle):
+                     uint16_t money, uint16_t ammo, float angle):
         type(type),
         position(position),
         id(id),
